Always fallback to limited range in y4m source
Also, change the default matrix.

There are too many limited range videos, and many video formats use it as default, so if a stream is full range, it will most likely be explicitly flagged as full range.
C420jpeg doesn't mean YUVJ420P (which is full range), it only means the chroma location is center.